Fix Refactor Redundant Method Call
In file: <original_filename>, uv_func a test expression has multiple calls of the same method. This is redundant and affects performance. On the other hand, if this is desired behavior in which multiple method calls with the same parameter returns different results, the logic should be reviewed to distinguish the method calls and make the code clean. iCR suggested that the method calls should be refactored and then the local variable should be used in the test expression.